Removed some unused exports, renamed the item modals, fixed formatting of TaskItem.
