--- scanhub-ui/src/utils/SizeVars.tsx
@@ -1,17 +1,0 @@
-/**
- * Copyright (C) 2024, BRAIN-LINK UG (haftungsbeschränkt). All Rights Reserved.
- * SPDX-License-Identifier: GPL-3.0-only OR LicenseRef-ScanHub-Commercial
- *
- * SizeVars.tsx contains defintiions of global variables.
- */
-
-export const navigation = {
-  height: '64px',
-}
-
-export const patientView = {
-  drawerWidth: '400px',
-  procedureListWidth: '300px',
-  iconButtonSize: '30px',
-  iconButtonVariant: 'soft',
-}
